// frontend/src/app/api/model/taskSort.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const TaskSort = {
    DueDate: 'DUE_DATE',
    CreatedAt: 'CREATED_AT',
} as const;
export type TaskSort = typeof TaskSort[keyof typeof TaskSort];